Extracted all attributes out to their own file
There were a number of attributes that were written in the same
file. Separating them out should help further refactoring and make
things more consistent
