+++ lib/json-schema/attributes/properties.rb
@@ -62,13 +62,5 @@ module JSON
         end
       end
     end
-
-    class PropertiesV4Attribute < PropertiesAttribute
-      # draft4 relies on its own RequiredAttribute validation at a higher level, rather than
-      # as an attribute of individual properties.
-      def self.required?(schema, options)
-        options[:strict] == true
-      end
-    end
   end
 end
